Moved MediaCaption to ImageBlock
