+++ src/plugins/image/ImageBlockStyle.js
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2016, Globo.com (https://github.com/globocom)
+ *
+ * License: MIT
+ */
+
+export default {
+  image: {
+    display: "block", // Eliminates whitespace between block and data fields block
+    maxWidth: "100%"
+  }
+};
