+++ src/styles/components/MediaStyle.js
@@ -18,11 +18,6 @@ export default {
     boxShadow: "0 1px 1px 0 #000",
     border: "solid 1px #ddd"
   },
-  block: {
-    display: "block", // Eliminates whitespace between block and data fields block
-    maxWidth: "100%",
-    userSelect: "none"
-  },
   selectedText: {
     display: "inline-block"
   }
